allow user pass in function to handle exchange ids(OpenMined#4909)
